[AN] refactor(#139): dto 이름 수정
